Fixed more warnings

+++ trunk/Bricksmith/Metal/LDraw/Renderer/LDrawDisplayList.m
@@ -506,7 +506,7 @@ static void immediateDraw(id<MTLRenderCommandEncoder>	renderEncoder,
 									  indexCount:tptr->line_count
 									   indexType:MTLIndexTypeUInt32
 									 indexBuffer:dl->indexBuffer
-							   indexBufferOffset:idx_null+tptr->line_off
+							   indexBufferOffset:(NSUInteger)(idx_null+tptr->line_off)
 								   instanceCount:1];
 
 		if(tptr->cond_line_count)
@@ -514,7 +514,7 @@ static void immediateDraw(id<MTLRenderCommandEncoder>	renderEncoder,
 									  indexCount:tptr->cond_line_count
 									   indexType:MTLIndexTypeUInt32
 									 indexBuffer:dl->indexBuffer
-							   indexBufferOffset:idx_null+tptr->cond_line_off
+							   indexBufferOffset:(NSUInteger)(idx_null+tptr->cond_line_off)
 								   instanceCount:1];
 		#else
 		if(tptr->line_count)
@@ -543,7 +543,7 @@ static void immediateDraw(id<MTLRenderCommandEncoder>	renderEncoder,
 										  indexCount:tptr->line_count
 										   indexType:MTLIndexTypeUInt32
 										 indexBuffer:dl->indexBuffer
-								   indexBufferOffset:idx_null+tptr->line_off
+								   indexBufferOffset:(NSUInteger)(idx_null+tptr->line_off)
 									   instanceCount:1];
 
 			if(tptr->tri_count)
@@ -551,7 +551,7 @@ static void immediateDraw(id<MTLRenderCommandEncoder>	renderEncoder,
 										  indexCount:tptr->tri_count
 										   indexType:MTLIndexTypeUInt32
 										 indexBuffer:dl->indexBuffer
-								   indexBufferOffset:idx_null+tptr->tri_off];
+								   indexBufferOffset:(NSUInteger)(idx_null+tptr->tri_off)];
 			#else
 			if(tptr->line_count)
 				[renderEncoder drawPrimitives:MTLPrimitiveTypeLine
@@ -1343,14 +1343,14 @@ void LDrawDLSessionDrawAndDestroy(id<MTLRenderCommandEncoder> renderEncoder, str
 												  indexCount:tptr->line_count
 												   indexType:MTLIndexTypeUInt32
 												 indexBuffer:dl->indexBuffer
-										   indexBufferOffset:idx_null+tptr->line_off];
+										   indexBufferOffset:(NSUInteger)(idx_null+tptr->line_off)];
 
 					if(tptr->tri_count)
 						[renderEncoder drawIndexedPrimitives:MTLPrimitiveTypeTriangle
 												  indexCount:tptr->tri_count
 												   indexType:MTLIndexTypeUInt32
 												 indexBuffer:dl->indexBuffer
-										   indexBufferOffset:idx_null+tptr->tri_off
+										   indexBufferOffset:(NSUInteger)(idx_null+tptr->tri_off)
 										   instanceCount:1];
 					#else
 					if(tptr->line_count)
@@ -1397,7 +1397,7 @@ void LDrawDLSessionDrawAndDestroy(id<MTLRenderCommandEncoder> renderEncoder, str
 			for(s = segments; s < cur_segment; ++s)
 			{
 				[renderEncoder setVertexBuffer:s->vertexBuffer offset:0 atIndex:BufferIndexInstanceInvariantData];
-				[renderEncoder setVertexBufferOffset:s->inst_base atIndex:BufferIndexPerInstanceData];
+				[renderEncoder setVertexBufferOffset:(NSUInteger)s->inst_base atIndex:BufferIndexPerInstanceData];
 
 				#if WANT_SMOOTH	
 				if(s->dl->line_count)
@@ -1405,7 +1405,7 @@ void LDrawDLSessionDrawAndDestroy(id<MTLRenderCommandEncoder> renderEncoder, str
 											  indexCount:s->dl->line_count
 											   indexType:MTLIndexTypeUInt32
 											 indexBuffer:s->indexBuffer
-									   indexBufferOffset:idx_null+s->dl->line_off
+									   indexBufferOffset:(NSUInteger)(idx_null+s->dl->line_off)
 										   instanceCount:s->inst_count];
 
 				if(s->dl->cond_line_count && s->is_wireframe)
@@ -1413,7 +1413,7 @@ void LDrawDLSessionDrawAndDestroy(id<MTLRenderCommandEncoder> renderEncoder, str
 											  indexCount:s->dl->cond_line_count
 											   indexType:MTLIndexTypeUInt32
 											 indexBuffer:s->indexBuffer
-									   indexBufferOffset:idx_null+s->dl->cond_line_off
+									   indexBufferOffset:(NSUInteger)(idx_null+s->dl->cond_line_off)
 										   instanceCount:s->inst_count];
 
 				if(s->dl->tri_count && !s->is_wireframe)
@@ -1421,7 +1421,7 @@ void LDrawDLSessionDrawAndDestroy(id<MTLRenderCommandEncoder> renderEncoder, str
 											  indexCount:s->dl->tri_count
 											   indexType:MTLIndexTypeUInt32
 											 indexBuffer:s->indexBuffer
-									   indexBufferOffset:idx_null+s->dl->tri_off
+									   indexBufferOffset:(NSUInteger)(idx_null+s->dl->tri_off)
 										   instanceCount:s->inst_count];
 				#else
 				if(s->dl->line_count)
@@ -1509,7 +1509,7 @@ void LDrawDLSessionDrawAndDestroy(id<MTLRenderCommandEncoder> renderEncoder, str
 											  indexCount:tptr->line_count
 											   indexType:MTLIndexTypeUInt32
 											 indexBuffer:dl->indexBuffer
-									   indexBufferOffset:idx_null+tptr->line_off
+									   indexBufferOffset:(NSUInteger)(idx_null+tptr->line_off)
 										   instanceCount:1];
 
 				if(tptr->tri_count)
@@ -1517,7 +1517,7 @@ void LDrawDLSessionDrawAndDestroy(id<MTLRenderCommandEncoder> renderEncoder, str
 											  indexCount:tptr->tri_count
 											   indexType:MTLIndexTypeUInt32
 											 indexBuffer:dl->indexBuffer
-									   indexBufferOffset:idx_null+tptr->tri_off
+									   indexBufferOffset:(NSUInteger)(idx_null+tptr->tri_off)
 										   instanceCount:1];
 				#else
 				if(tptr->line_count)
